22.04.22 09:00

+++ src/main/java/ru/kata/spring/boot_security/demo/model/User.java
@@ -6,7 +6,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import javax.persistence.*;
 import java.util.Collection;
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 
 @Entity
@@ -16,13 +15,13 @@ public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
 
     @Column(name = "name")
     private String name;
 
     @Column(name = "last_name")
-    private String lastName;
+        private String lastName;
 
     @Column(name = "email", unique = true, length = 100)
     private String email;
@@ -46,11 +45,11 @@ public class User implements UserDetails {
         this.password = password;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
